fix(oxlint): remove .d.ts module augmentations per review feedback
Extract mock helper functions and fixture data to separate test utility
files (gitMock.ts, growthbookMock.tsx) outside __mocks__/ directories.
The __mocks__/ files now import from these helpers and re-export module
exports for vi.mock() auto-resolution. Test files import helpers directly
from the utility files instead of from __mocks__/ paths or real module
names with .d.ts augmentation.

This addresses the reviewer's objection to augmenting every mocked
module's types. No module augmentation, no type casting, no __mocks__/
imports in test files.
