+++ packages/changesets-renovate/__mocks__/simple-git.ts
@@ -1,18 +1,4 @@
-import { vi } from 'vitest'
-
-export const defaultGitValues = {
-  add: () => undefined,
-  branch: () => ({
-    current: '',
-  }),
-  commit: () => undefined,
-  diffSummary: () => ({}),
-  push: () => undefined,
-  revparse: () => undefined,
-  show: () => '',
-}
-
-export const mockSimpleGit = vi.fn(() => defaultGitValues)
+import { mockSimpleGit } from '../src/__tests__/gitMock.js'
 
 const simpleGit = () => mockSimpleGit()
